+++ typer/__init__.py
@@ -26,8 +26,8 @@ from click.utils import get_text_stream as get_text_stream
 from click.utils import open_file as open_file
 
 from . import colors as colors
-from .launch import launch as launch
 from .main import Typer as Typer
+from .main import launch as launch
 from .main import run as run
 from .models import CallbackParam as CallbackParam
 from .models import Context as Context
